dev: Connected to gcloud datastore
The app now is able to take registered users and save them to the google data store. There is a error with the login. Passport Js local strategy is working. Compare password works. There is a unhandled error if no user is found in the dtabase that needs to be fixed. Also there is a error that is preventing the user to be serialized into a session.