dev: Added React Dropzone for mp3 file upload. Also built mp3 info form .
The file upload is not connect to database and redux at this time, setup form to render per each file uploaded after upload is completed.  Validate that only Mp3 are being upladed to file uploader. Verify each file is being uploaded with the user who uploaded id in order to be pulled into the list.